Separate library settings
